Add "package download" pattern, which provides a submit button to download a PDF package.
