+++ packages/forms/src/components.ts
@@ -8,6 +8,12 @@ import {
 import { type FormSession, nullSession, sessionIsComplete } from './session.js';
 import { type ActionName } from './submission.js';
 
+export type PackageDownloadProps = PatternProps<{
+  type: 'package-download';
+  text: string;
+  actions: PromptAction[];
+}>;
+
 export type TextInputProps = PatternProps<{
   type: 'input';
   inputId: string;
