+++ packages/forms/src/documents/pdf/generate.ts
@@ -15,6 +15,7 @@ export const createFormOutputFieldData = (
       return;
     }
     const outputFieldId = output.formFields[patternId];
+    console.log('***', patternId, docField, outputFieldId);
     if (outputFieldId === '') {
       console.error(`empty outputFieldId for field: ${patternId}: ${docField}`);
       return;
@@ -65,7 +66,7 @@ export const fillPDF = async (
       .sort((a, b) => a.name.localeCompare(b.name));
 
     // Console log the resulting array
-    console.log('uniqueNamesArray:', uniqueNamesArray);
+    //console.log('uniqueNamesArray:', uniqueNamesArray);
 
     // fields.map(field => {
     //   console.log('field name is:', field.getName());
