+++ packages/forms/src/patterns/package-download/index.ts
@@ -1,0 +1,44 @@
+import * as z from 'zod';
+
+import { type Pattern, type PatternConfig } from '../../pattern.js';
+import { type PackageDownloadProps } from '../../components.js';
+import { type ActionName, getActionString } from '../../submission.js';
+import { safeZodParseFormErrors } from '../../util/zod.js';
+
+const configSchema = z.object({
+  text: z.string().min(1),
+});
+export type PackageDownloadPattern = Pattern<z.infer<typeof configSchema>>;
+
+export const packageDownloadConfig: PatternConfig<PackageDownloadPattern> = {
+  displayName: 'Package download',
+  iconPath: 'block-icon.svg',
+  initial: {
+    text: 'Description text...',
+  },
+  parseConfigData: obj => safeZodParseFormErrors(configSchema, obj),
+  getChildren() {
+    return [];
+  },
+  createPrompt(_, session, pattern, options) {
+    const actionName: ActionName = getActionString({
+      handlerId: 'page-set',
+      patternId: pattern.id,
+    });
+    return {
+      props: {
+        _patternId: pattern.id,
+        type: 'package-download' as const,
+        actions: [
+          {
+            type: 'submit',
+            submitAction: actionName,
+            text: 'Download PDF',
+          },
+        ],
+        text: pattern.data.text,
+      } as PackageDownloadProps,
+      children: [],
+    };
+  },
+};
